third_party/WebKit/LayoutTests/http/tests/inspector/security/mixed-content-and-subresources-with-cert-errors-expected.txt

Tests addition of explanation when there was both mixed content and subresources were loaded with certificate errors.

<DIV class=security-explanation security-explanation-insecure >
    <DIV class=security-property security-property-insecure >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Active Mixed Content
        </DIV>
        <DIV >
You have recently allowed non-secure content (such as scripts or iframes) to run on this site.
        </DIV>
        <DIV class=security-mixed-content link >
View requests in Network Panel
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Mixed Content
        </DIV>
        <DIV >
The site includes HTTP resources.
        </DIV>
        <DIV class=security-mixed-content link >
View requests in Network Panel
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-insecure >
    <DIV class=security-property security-property-insecure >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Active content with certificate errors
        </DIV>
        <DIV >
You have recently allowed content loaded with certificate errors (such as scripts or iframes) to run on this site.
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-neutral >
    <DIV class=security-property security-property-neutral >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Content with certificate errors
        </DIV>
        <DIV >
This site includes resources that were loaded with certificate errors.
        </DIV>
    </DIV>
</DIV>
